Allow manual install to be requested for NCC installs
If an NCC install goes bad, there was no method for the user to
force a manual install.  This adds a button to request a manual
install to be requested by MO.

The dialog was copied and modified from the quick install plugin
so there may be leftover stuff related to that.  This also slightly
changes the look of the dialog that would've popped up.